chore (refs T22614): move DpAccordion to client/js.
