+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/bundles/administrationUnregisteredPublicagencyEmail.js
@@ -10,7 +10,7 @@
 /**
  * This is the entrypoint for administration_unregistered_publicagency_email.html.twig
  */
-import DpAccordion from '@DemosPlanCoreBundle/components/DpAccordion'
+import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpTiptap from '@DemosPlanCoreBundle/components/DpTiptap'
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
