+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/DpSimplifiedNewStatementForm.vue
@@ -349,7 +349,7 @@
 
 <script>
 import { DpInput, DpLabel } from 'demosplan-ui/components'
-import DpAccordion from '@DemosPlanCoreBundle/components/DpAccordion'
+import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpButtonRow from '@DpJs/components/core/DpButtonRow'
 import DpDatepicker from '@DpJs/components/core/form/DpDatepicker'
 import DpMultiselect from '@DpJs/components/core/form/DpMultiselect'
