+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/admin/AuthorizedUsersList.vue
@@ -291,7 +291,7 @@
 <script>
 import { dpApi, dpRpc } from '@DemosPlanCoreBundle/plugins/DpApi'
 import { DpButton, DpInput } from 'demosplan-ui/components'
-import DpAccordion from '@DemosPlanCoreBundle/components/DpAccordion'
+import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpButtonRow from '@DpJs/components/core/DpButtonRow'
 import DpDataTableExtended from '@DemosPlanCoreBundle/components/DpDataTable/DpDataTableExtended'
 import DpTextArea from '@DpJs/components/core/form/DpTextArea'
