+++ demosplan/DemosPlanStatementBundle/Resources/client/js/bundles/viewStatement.js
@@ -12,7 +12,7 @@
  */
 
 import AnimateById from '@DpJs/lib/AnimateById'
-import DpAccordion from '@DemosPlanCoreBundle/components/DpAccordion'
+import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpHeightLimit from '@DpJs/components/core/HeightLimit'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
