+++ demosplan/DemosPlanUserBundle/Resources/client/js/bundles/developmentReleaseStory.js
@@ -11,7 +11,7 @@
  * This is the entrypoint for development_release_story_threadentry_list.html.twig
  */
 
-import DpAccordion from '@DemosPlanCoreBundle/components/DpAccordion'
+import DpAccordion from '@DpJs/components/core/DpAccordion'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import { VPopover } from 'demosplan-ui/directives'
