+++ demosplan/DemosPlanUserBundle/Resources/client/js/bundles/editOrga.js
@@ -11,7 +11,7 @@
  * This is the entrypoint for edit_orga.html.twig
  */
 
-import DpAccordion from '@DemosPlanCoreBundle/components/DpAccordion'
+import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpTiptap from '@DemosPlanCoreBundle/components/DpTiptap'
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
