+++ demosplan/DemosPlanUserBundle/Resources/client/js/components/DpCreateItem.vue
@@ -46,7 +46,7 @@
 
 <script>
 import { mapActions, mapState } from 'vuex'
-import DpAccordion from '@DemosPlanCoreBundle/components/DpAccordion'
+import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpButtonRow from '@DpJs/components/core/DpButtonRow'
 import dpValidateMixin from '@DpJs/lib/validation/dpValidateMixin'
